+++ src/test/java/tools/jackson/databind/tofix/MapEntryFormat1419Test.java
@@ -29,7 +29,7 @@ class MapEntryFormat1419Test extends DatabindTestUtil {
         }
     }
 
-    private final ObjectMapper MAPPER = new ObjectMapper();
+    private final ObjectMapper MAPPER = newJsonMapper();
 
     @JacksonTestFailureExpected
     @Test
